rendering WeatherPage but not able to pass function as props down to WeatherPage
